Fixed "text-shadow"
The "text-shadow" is used to add shadow to text.
Here in 
```
text-shadow: 1px 1px 1px rgba(255, 255, 255, 0.2);
```
It means that, the text has shadow of 1px in x(horizontal) and 1px in y(vertical) and has 1px blur-radius and rgb is for the color, 0 is for pure back and 255 is for white. and a is for alpha channel,where 0.0 is fully transparent and 1.0 fully opaque.